+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/AppCenter.java
@@ -289,6 +289,15 @@ public class AppCenter {
     }
 
     /**
+     * Set the two-letter ISO country code to send to the backend.
+     *
+     * @param countryCode the two-letter ISO country code. See <code>https://www.iso.org/obp/ui/#search</code> for more information.
+     */
+    public static void setCountryCode(String countryCode) {
+        DeviceInfoHelper.setCountryCode(countryCode);
+    }
+
+    /**
      * Get the current version of App Center SDK.
      *
      * @return The current version of App Center SDK.